refactor: move describeProduct to it's own module
